Remove tags that can be implied
The tags <html>, <body> and <head> are implied when omitted. <html> is implied as the outermost tag if omitted, but we don't omit it because of the lang= attribute. <head> is implied as the first tag in <html> and the first tag that does not fit in <head> (<h1> in our case) implies <body>. Closing tags are also implied.

Since this is a base html, it makes sense to omit tags that can be implied. By not using a <body> tag, we also shave off one layer of indentation.